Fix integer types

+++ src/rest/creds.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct ManagerCreds {
     pub password: String,
-    pub login: i32,
+    pub login: i64,
 }
